extract index structure intto a serprate file
