+++ dpp/src/util/mod.rs
@@ -2,6 +2,7 @@ pub mod cbor_value;
 pub mod deserializer;
 pub mod entropy_generator;
 pub mod hash;
+pub mod index;
 pub mod json_path;
 pub mod json_schema;
 pub mod json_value;
